+++ test/src/unit-capi-config.cc
@@ -59,7 +59,7 @@ void check_load_correct_file() {
   ofs << "   # comment line\n";
   ofs << "sm.tile_cache_size 1000\n";
   ofs << "# another comment line\n";
-  ofs << "sm.num_async_threads 10 # some comment\n";
+  ofs << "sm.consolidation.steps 2 # some comment\n";
   ofs << "#    last comment line\n";
   ofs.close();
 
@@ -112,7 +112,7 @@ void check_load_incorrect_file_missing_value() {
   ofs << "   # comment line\n";
   ofs << "sm.tile_cache_size    \n";
   ofs << "# another comment line\n";
-  ofs << "sm.num_async_threads 10\n";
+  ofs << "sm.consolidation.steps 2 # some comment\n";
   ofs << "#    last comment line\n";
   ofs.close();
 
@@ -142,7 +142,7 @@ void check_load_incorrect_file_extra_word() {
   ofs << "   # comment line\n";
   ofs << "sm.tile_cache_size 1000\n";
   ofs << "# another comment line\n";
-  ofs << "sm.num_async_threads 10 some comment\n";
+  ofs << "sm.consolidation.steps 2 some comment\n";
   ofs << "#    last comment line\n";
   ofs.close();
 
@@ -232,11 +232,11 @@ void check_save_to_file() {
   ss << "sm.memory_budget 5368709120\n";
   ss << "sm.memory_budget_var 10737418240\n";
   ss << "sm.num_tbb_threads -1\n";
-  ss << "sm.offsets_format bytes\n";
   ss << "sm.skip_checksum_validation false\n";
   ss << "sm.sub_partitioner_memory_budget 0\n";
   ss << "sm.tile_cache_size 10000000\n";
   ss << "sm.vacuum.mode fragments\n";
+  ss << "sm.var_offsets.mode bytes\n";
   ss << "vfs.azure.block_list_block_size 5242880\n";
   ss << "vfs.azure.max_parallel_ops " << std::thread::hardware_concurrency()
      << "\n";
@@ -442,7 +442,7 @@ TEST_CASE("C API: Test config iter", "[capi], [config]") {
   rc = tiledb_config_set(config, "vfs.hdfs.username", "stavros", &error);
   CHECK(rc == TILEDB_OK);
   CHECK(error == nullptr);
-  rc = tiledb_config_set(config, "sm.offsets_format", "elements", &error);
+  rc = tiledb_config_set(config, "sm.var_offsets.mode", "elements", &error);
   CHECK(rc == TILEDB_OK);
   CHECK(error == nullptr);
 
@@ -477,7 +477,7 @@ TEST_CASE("C API: Test config iter", "[capi], [config]") {
   all_param_values["sm.consolidation.step_size_ratio"] = "0.0";
   all_param_values["sm.consolidation.mode"] = "fragments";
   all_param_values["sm.vacuum.mode"] = "fragments";
-  all_param_values["sm.offsets_format"] = "elements";
+  all_param_values["sm.var_offsets.mode"] = "elements";
 
   all_param_values["vfs.min_batch_gap"] = "512000";
   all_param_values["vfs.min_batch_size"] = "20971520";