+++ tiledb/sm/config/config.cc
@@ -82,7 +82,7 @@ const std::string Config::SM_CONSOLIDATION_STEP_MAX_FRAGS = "4294967295";
 const std::string Config::SM_CONSOLIDATION_STEP_SIZE_RATIO = "0.0";
 const std::string Config::SM_CONSOLIDATION_MODE = "fragments";
 const std::string Config::SM_VACUUM_MODE = "fragments";
-const std::string Config::SM_OFFSETS_FORMAT = "bytes";
+const std::string Config::SM_OFFSETS_FORMAT_MODE = "bytes";
 const std::string Config::VFS_MIN_PARALLEL_SIZE = "10485760";
 const std::string Config::VFS_MIN_BATCH_GAP = "512000";
 const std::string Config::VFS_MIN_BATCH_SIZE = "20971520";
@@ -198,7 +198,7 @@ Config::Config() {
   param_values_["sm.consolidation.steps"] = SM_CONSOLIDATION_STEPS;
   param_values_["sm.consolidation.mode"] = SM_CONSOLIDATION_MODE;
   param_values_["sm.vacuum.mode"] = SM_VACUUM_MODE;
-  param_values_["sm.offsets_format"] = SM_OFFSETS_FORMAT;
+  param_values_["sm.var_offsets.mode"] = SM_OFFSETS_FORMAT_MODE;
   param_values_["vfs.min_parallel_size"] = VFS_MIN_PARALLEL_SIZE;
   param_values_["vfs.min_batch_gap"] = VFS_MIN_BATCH_GAP;
   param_values_["vfs.min_batch_size"] = VFS_MIN_BATCH_SIZE;
@@ -403,7 +403,7 @@ Status Config::unset(const std::string& param) {
     param_values_["sm.memory_budget"] = SM_MEMORY_BUDGET;
   } else if (param == "sm.memory_budget_var") {
     param_values_["sm.memory_budget_var"] = SM_MEMORY_BUDGET_VAR;
-  } else if (param == "sm.memory_budget") {
+  } else if (param == "sm.sub_partitioner_memory_budget") {
     param_values_["sm.sub_partitioner_memory_budget"] =
         SM_SUB_PARTITIONER_MEMORY_BUDGET;
   } else if (param == "sm.enable_signal_handlers") {
@@ -436,8 +436,8 @@ Status Config::unset(const std::string& param) {
     param_values_["sm.consolidation.mode"] = SM_CONSOLIDATION_MODE;
   } else if (param == "sm.vacuum.mode") {
     param_values_["sm.vacuum.mode"] = SM_VACUUM_MODE;
-  } else if (param == "sm.offsets_format") {
-    param_values_["sm.offsets_format"] = SM_OFFSETS_FORMAT;
+  } else if (param == "sm.var_offsets.mode") {
+    param_values_["sm.var_offsets.mode"] = SM_OFFSETS_FORMAT_MODE;
   } else if (param == "vfs.min_parallel_size") {
     param_values_["vfs.min_parallel_size"] = VFS_MIN_PARALLEL_SIZE;
   } else if (param == "vfs.min_batch_gap") {
@@ -541,7 +541,7 @@ Status Config::unset(const std::string& param) {
     param_values_["vfs.s3.proxy_password"] = VFS_S3_PROXY_PASSWORD;
   } else if (param == "vfs.s3.verify_ssl") {
     param_values_["vfs.s3.verify_ssl"] = VFS_S3_VERIFY_SSL;
-  } else if (param == "vfs.hdfs.name_node") {
+  } else if (param == "vfs.hdfs.name_node_uri") {
     param_values_["vfs.hdfs.name_node_uri"] = VFS_HDFS_NAME_NODE_URI;
   } else if (param == "vfs.hdfs.username") {
     param_values_["vfs.hdfs.username"] = VFS_HDFS_USERNAME;
@@ -632,7 +632,7 @@ Status Config::sanity_check(
     RETURN_NOT_OK(utils::parse::convert(value, &v32));
   } else if (param == "sm.consolidation.step_size_ratio") {
     RETURN_NOT_OK(utils::parse::convert(value, &vf));
-  } else if (param == "sm.offsets_format") {
+  } else if (param == "sm.var_offsets.mode") {
     if (value != "bytes" && value != "elements")
       return LOG_STATUS(
           Status::ConfigError("Invalid offsets format parameter value"));